[FIX] access and dependencies
