+++ spp_area/models/__init__.py
@@ -2,3 +2,5 @@
 
 
 from . import registrant
+from . import area
+from . import user
