+++ spp_registry_base/__manifest__.py
@@ -17,6 +17,7 @@
         "g2p_registry_individual",
         "g2p_registry_group",
         "g2p_registry_membership",
+        "spp_user_roles",
     ],
     "data": [
         "security/security_access.xml",
