+++ spp_user_roles/models/__init__.py
@@ -1,6 +1,3 @@
 # Part of OpenSPP. See LICENSE file for full copyright and licensing details.
 
 from . import role
-from . import user
-from . import area
-from . import res_partner
