+++ spp_user_roles/tests/__init__.py
@@ -1,2 +1,2 @@
-from . import test_res_partner
-from . import test_user
+# from . import test_res_partner
+# from . import test_user
